# Fix Directory.EnumerateFiles regression