--- a/src/EventStore.Core/TransactionLog/Chunks/TFChunkDbConfig.cs
+++ b/src/EventStore.Core/TransactionLog/Chunks/TFChunkDbConfig.cs
@@ -15,7 +15,7 @@ namespace EventStore.Core.TransactionLog.Chunks {
 		public readonly ICheckpoint ReplicationCheckpoint;
 		public readonly ICheckpoint IndexCheckpoint;
 		public readonly ICheckpoint StreamExistenceFilterCheckpoint;
-		public readonly IFileNamingStrategy FileNamingStrategy;
+		public readonly IVersionedFileNamingStrategy FileNamingStrategy;
 		public readonly bool InMemDb;
 		public readonly bool Unbuffered;
 		public readonly bool WriteThrough;
@@ -26,7 +26,7 @@ namespace EventStore.Core.TransactionLog.Chunks {
 		public readonly long MaxTruncation;
 
 		public TFChunkDbConfig(string path,
-			IFileNamingStrategy fileNamingStrategy,
+			IVersionedFileNamingStrategy fileNamingStrategy,
 			int chunkSize,
 			long maxChunksCacheSize,
 			ICheckpoint writerCheckpoint,