--- a/src/EventStore.Core/TransactionLog/FileNamingStrategy/IVersionedFileNamingStrategy.cs
+++ b/src/EventStore.Core/TransactionLog/FileNamingStrategy/IVersionedFileNamingStrategy.cs
@@ -1,11 +1,12 @@
 namespace EventStore.Core.TransactionLog.FileNamingStrategy {
-	public interface IFileNamingStrategy {
+	public interface IVersionedFileNamingStrategy {
 		string GetFilenameFor(int index, int version);
 		string DetermineBestVersionFilenameFor(int index);
 		string[] GetAllVersionsFor(int index);
 		string[] GetAllPresentFiles();
-
 		string GetTempFilename();
 		string[] GetAllTempFiles();
+		int GetIndexFor(string fileName);
+		int GetVersionFor(string fileName);
 	}
 }