… src/EventStore.Core/TransactionLog/Chunks/TFChunkDbTruncator.cs
… src/EventStore.Core/TransactionLog/FileNamingStrategy/VersionedPatternFileNamingStrategy.cs