Fixed: Directory.EnumerateFiles regression causing slower startup/truncation times on large databases

Multiple functions namely `GetAllLatestChunkVersions()`, `EnsureNoExcessiveChunks()`, `RemoveOldChunksVersions()` and `TruncateDb()` have a computational complexity of O(n^2) where n = number of chunks.

This was not noticeable earlier since Directory.EnumerateFiles was quite fast but with this ~50x regression in Directory.EnumerateFiles: https://github.com/dotnet/runtime/issues/31214, large databases can take several minutes to load compared to a few seconds in the past.

This PR fixes it by reducing the computational complexity of these methods to O(n lg n).
Existing tests already cover most of these changes - some tests have been strengthened to cover more cases.

## Benchmarks
- Tested with `--skip-index-verify --skip-db-verify` on Ubuntu 20.04/Windows Server 2019 with an AWS EBS gp2 volume on a t2.medium VM. v5-master was tested on Ubuntu 18.04 since I later realized mono runtime we need wasn't available for 20.04.
- Chunk size was adjusted to 256 KB instead of 256 MB
- On linux page cache was cleared between tests with `echo 3 > /proc/sys/vm/drop_caches`. Page cache wasn't cleared on windows.
- Truncation tests were done by simply copying `writer.chk` to `truncate.chk` and starting the database
- the min/max in the current benchmark refers to the time taken without and with the commit: a0940fd7e6c4c7cc50d7b0831bbdc106597761b3. It's basically the difference between loading a fully non-scavenged vs fully scavenged database.

| Test                                              | v5-master (linux)     | v5-master (win) | master (linux) | master (win) | this branch (linux)             | this branch (win)               |
|---------------------------------------------------|-----------------------|-----------------|----------------|--------------|---------------------------------|---------------------------------|
| Opening of chunks (10k 256kb chunks)              | 180 secs              | 7 secs          | 116 secs       | 85 secs      | min: 16 secs / max: 16 secs     | min: 19 secs / max: 25 secs     |
| Opening of chunks (80k 256kb chunks)              | OOM (4GB) & very slow | 4 mins          | 114 mins       | 52 mins      | min: 2.3 mins / max: 3.5 mins   | min: 2 mins / max: 4.6 mins     |
| Check excessive chunks (10k 256kb chunks)         | 101 secs              | 800 ms          | 67 secs        | 53 secs      | 300 ms                          | 700 ms                          |
| Check excessive chunks (80k 256kb chunks)         | Very slow             | 7 secs          | 113 mins       | 74 mins      | 1.5 secs                        | 5 secs                          |
| Remove old chunks (10k 256kb chunks)              | 101 secs              | 600 ms          | 121 secs       | 56 secs      | 250 ms                          | 700 ms                          |
| Remove old chunks (80k 256kb chunks)              | Very slow             | 4.7 secs        | 112 mins       | 104 mins     | 1.7 secs                        | 5 secs                          |
| Total opening time, excl index (10k 256kb chunks) | 6.4 mins              | 8.5 secs        | 5 mins         | 3 mins       | min: 16.5 secs / max: 16.5 secs | min: 20.5 secs / max: 26.5 secs |
| Total opening time, excl index (80k 256kb chunks) | Very slow             | 4.2 mins        | 5.5 hours      | 3.8 hours    | min: 2.3 mins / max: 3.5 mins   | min: 2.2 mins / max: 4.8 mins   |
| Truncate DB (10k 256kb chunks)                    | 115 secs              | 1.6 secs        | 80 secs        | 54 secs      | min: 130 ms / max: 9 secs       | min: 150 ms / max: 22 secs      |
| Truncate DB (80k 256kb chunks)                    | Very slow             | 2.5 mins        | 61 mins        | 52 mins      | min: 1.1s / max: 69 secs        | min: 1s / max: 4.3 mins         |

## Summary
The massive performance boost in loading times is applicable to:
- Users running on linux with any version (assuming that < v5-linux is as slow as v5-linux or slower)
- Users running on windows with v20 or later (v5-win users will not see a significant performance difference)