@tambeau 2022-01-13
Related Support Ticket https://eventstore.freshdesk.com/a/tickets/110065