@shaan1337 2022-01-17
Latest benchmarks on 80,000 256kb chunks database

## Linux / non-scavenged chunks:

Opening: 2 mins 2 secs
Ensure no excessive chunks: 150 ms
Remove old chunks: 150 ms

## Linux / scavenged chunks (simulated by removing version = 0 optimization):

Opening: 2 mins 13 secs
Ensure no excessive chunks: 180 ms
Remove old chunks: 150 ms

## Windows / non-scavenged chunks:

Opening: 2 mins 10 secs
Ensure no excessive chunks: 160 ms
Remove old chunks: 120 ms

## Windows / scavenged chunks (simulated by removing version = 0 optimization):

Opening: 3 mins 58 secs
Ensure no excessive chunks: 185 ms
Remove old chunks: 200 ms